+++ src/Threax.Home.Core/ICoreSensor.cs
@@ -19,19 +19,19 @@ namespace Threax.Home.Core
         /// </summary>
         String Id { get; set; }
 
-        float? TempValue { get; set; }
+        double? TempValue { get; set; }
 
         Units? TempUnits { get; set; }
 
-        float? LightValue { get; set; }
+        double? LightValue { get; set; }
 
         Units? LightUnits { get; set; }
 
-        float? HumidityValue { get; set; }
+        double? HumidityValue { get; set; }
 
         Units? HumidityUnits { get; set; }
 
-        float? UvValue { get; set; }
+        double? UvValue { get; set; }
 
         Units? UvUnits { get; set; }
     }
